+++ src/app/controllers/index.ts
@@ -4,3 +4,4 @@ export * from './file.controller';
 export * from './list-files.controller';
 export * from './orm.controller';
 export * from './terminal.controller';
+export * from './transform.controller';